Move xfel.mono_simulation.max_like to serialtbx
